pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "day-trade-intelligence"
version = "0.1.0"
description = "Production-ready day trading orchestration platform using Alpaca data and IBKR execution."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "alpaca-py>=0.42.0",
    "celery[redis]>=5.4.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "ib-insync>=0.9.86",
    "pydantic-settings>=2.7.0",
    "psycopg[binary]>=3.2.0",
    "python-json-logger>=2.0.7",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
